docs: ✏️ fix doc warnings for multiSig + StatType
cleans up doc gen warnings around multiSig references and StatType not
being exposed in the docs
