+++ src/api/procedures/__tests__/addTransferRestriction.ts
@@ -25,11 +25,12 @@ import {
   CountryCode,
   ErrorCode,
   StatJurisdictionClaimInput,
+  StatType,
   TransferRestriction,
   TransferRestrictionType,
   TxTags,
 } from '~/types';
-import { PolymeshTx, StatType, TickerKey } from '~/types/internal';
+import { PolymeshTx, TickerKey } from '~/types/internal';
 import * as utilsConversionModule from '~/utils/conversion';
 
 jest.mock(
