+++ src/api/procedures/__tests__/removeAssetStat.ts
@@ -20,9 +20,10 @@ import {
   ErrorCode,
   RemoveAssetStatParams,
   StatClaimType,
+  StatType,
   TxTags,
 } from '~/types';
-import { PolymeshTx, StatType, TickerKey } from '~/types/internal';
+import { PolymeshTx, TickerKey } from '~/types/internal';
 import * as utilsConversionModule from '~/utils/conversion';
 
 jest.mock(
