+++ src/api/procedures/__tests__/setTransferRestrictions.ts
@@ -27,11 +27,12 @@ import {
   ClaimType,
   ErrorCode,
   Identity,
+  StatType,
   TransferRestriction,
   TransferRestrictionType,
   TxTags,
 } from '~/types';
-import { PolymeshTx, StatType, TickerKey } from '~/types/internal';
+import { PolymeshTx, TickerKey } from '~/types/internal';
 import * as utilsConversionModule from '~/utils/conversion';
 
 jest.mock(
