+++ src/testUtils/mocks/dataSources.ts
@@ -195,11 +195,12 @@ import {
   ProtocolFees,
   ResultSet,
   SignerType,
+  StatType,
   SubsidyWithAllowance,
   TxTags,
   UnsubCallback,
 } from '~/types';
-import { Consts, Extrinsics, GraphqlQuery, PolymeshTx, Queries, StatType } from '~/types/internal';
+import { Consts, Extrinsics, GraphqlQuery, PolymeshTx, Queries } from '~/types/internal';
 import { ArgsType, Mutable, tuple } from '~/types/utils';
 import { STATE_RUNTIME_VERSION_CALL, SYSTEM_VERSION_RPC_CALL } from '~/utils/constants';
