+++ src/types/index.ts
@@ -30,7 +30,6 @@ import {
   PolymeshTransaction,
   PolymeshTransactionBatch,
 } from '~/internal';
-import { StatType } from '~/types/internal';
 import { Modify } from '~/types/utils';
 
 export * from '~/generated/types';
@@ -331,6 +330,23 @@ export type InputStatType =
       claimIssuer: StatClaimIssuer;
     };
 
+/**
+ * Represents the StatType from the `statistics` module.
+ *
+ * @note the chain doesn't use "Scoped" types, but they are needed here to discriminate the input instead of having an optional input
+ */
+export enum StatType {
+  Count = 'Count',
+  Balance = 'Balance',
+  /**
+   * ScopedCount is an SDK only type, on chain it is `Count` with a claimType option present
+   */
+  ScopedCount = 'ScopedCount',
+  /**
+   * ScopedPercentage is an SDK only type, on chain it is `Balance` with a claimType option present
+   */
+  ScopedBalance = 'ScopedBalance',
+}
 export interface IdentityWithClaims {
   identity: Identity;
   claims: ClaimData[];
